#69 - Voters

See merge request !116

+++ src/AppBundle/Controller/API/ProjectCostTypeController.php
@@ -4,6 +4,7 @@ namespace AppBundle\Controller\API;
 
 use AppBundle\Entity\ProjectCostType;
 use AppBundle\Form\ProjectCostType\CreateType;
+use AppBundle\Security\ProjectVoter;
 use MainBundle\Controller\API\ApiController;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
@@ -78,6 +79,12 @@ class ProjectCostTypeController extends ApiController
      */
     public function getAction(ProjectCostType $projectCostType)
     {
+        $project = $projectCostType->getProject();
+        if (!$project) {
+            throw new \LogicException('Project does not exist!');
+        }
+        $this->denyAccessUnlessGranted(ProjectVoter::VIEW, $project);
+
         return $this->createApiResponse($projectCostType);
     }
 
@@ -94,6 +101,12 @@ class ProjectCostTypeController extends ApiController
      */
     public function editAction(Request $request, ProjectCostType $projectCostType)
     {
+        $project = $projectCostType->getProject();
+        if (!$project) {
+            throw new \LogicException('Project does not exist!');
+        }
+        $this->denyAccessUnlessGranted(ProjectVoter::EDIT, $project);
+
         $data = $request->request->all();
         $form = $this->createForm(CreateType::class, $projectCostType, ['csrf_protection' => false]);
         $form->submit($data, false);
@@ -128,6 +141,12 @@ class ProjectCostTypeController extends ApiController
      */
     public function deleteAction(ProjectCostType $projectCostType)
     {
+        $project = $projectCostType->getProject();
+        if (!$project) {
+            throw new \LogicException('Project does not exist!');
+        }
+        $this->denyAccessUnlessGranted(ProjectVoter::DELETE, $project);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($projectCostType);
         $em->flush();
